Fixed element resize event issue

+++ src/Blazor.Components.Common.JsInterop/Resize/ResizeEventArgs.cs
@@ -10,11 +10,11 @@ namespace Blazor.Components.Common.JsInterop.GlobalMouseEvents
 		/// <summary>
 		/// Element or Window height
 		/// </summary>
-		public int Height { get; set; }
+		public double Height { get; set; }
 		/// <summary>
 		/// Element or Window width
 		/// </summary>
-		public int Width { get; set; }
+		public double Width { get; set; }
 		
 		/// <summary>
 		/// Registered event id
